Fix names of existing tests
