+++ extensions/resteasy-reactive/rest/deployment/src/test/java/io/quarkus/resteasy/reactive/server/test/CancellableUniTest.java
@@ -3,15 +3,11 @@ package io.quarkus.resteasy.reactive.server.test;
 import static io.restassured.RestAssured.when;
 import static org.awaitility.Awaitility.await;
 import static org.hamcrest.CoreMatchers.equalTo;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.*;
 
 import java.net.URL;
 import java.time.Duration;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionStage;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Supplier;
 
 import jakarta.inject.Inject;
 import jakarta.ws.rs.GET;
@@ -19,16 +15,18 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 
+import org.jboss.resteasy.reactive.server.Cancellable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import io.quarkus.test.QuarkusExtensionTest;
 import io.quarkus.test.common.http.TestHTTPResource;
+import io.smallrye.mutiny.Uni;
 import io.vertx.core.Vertx;
 import io.vertx.ext.web.client.WebClient;
 
-public class CancelableCompletionStageTest {
+public class CancellableUniTest {
 
     @RegisterExtension
     static QuarkusExtensionTest runner = new QuarkusExtensionTest()
@@ -47,17 +45,35 @@ public class CancelableCompletionStageTest {
 
     @Test
     public void testNormal() {
-        when().get("test")
+        doTestNormal("1");
+    }
+
+    @Test
+    public void testDefaultCancellable() {
+        doTestCancel("1", Resource.COUNT, 1);
+    }
+
+    @Test
+    public void testUnCancellable() {
+        doTestCancel("2", Resource.COUNT, 2);
+    }
+
+    @Test
+    public void testCancellable() {
+        doTestCancel("3", Resource.COUNT, 1);
+    }
+
+    private void doTestNormal(String path) {
+        when().get("test/" + path)
                 .then()
                 .statusCode(200)
                 .body(equalTo("Hello, world"));
     }
 
-    @Test
-    public void testCancel() {
+    private void doTestCancel(String path, AtomicInteger count, int expected) {
         WebClient client = WebClient.create(vertx);
 
-        client.get(url.getPort(), url.getHost(), "/test").send();
+        client.get(url.getPort(), url.getHost(), "/test/" + path).send();
 
         try {
             // make sure we did make the proper request
@@ -70,7 +86,7 @@ public class CancelableCompletionStageTest {
             Thread.sleep(7_000);
 
             // if the count did not increase, it means that Uni was cancelled
-            assertEquals(1, Resource.COUNT.get());
+            assertEquals(expected, count.get());
         } catch (InterruptedException ignored) {
 
         } finally {
@@ -80,7 +96,6 @@ public class CancelableCompletionStageTest {
 
             }
         }
-
     }
 
     @Path("test")
@@ -90,17 +105,31 @@ public class CancelableCompletionStageTest {
 
         @GET
         @Produces(MediaType.TEXT_PLAIN)
-        public CompletionStage<String> hello() {
+        @Path("1")
+        public Uni<String> defaultCancelableHello() {
             COUNT.incrementAndGet();
-            return CompletableFuture.supplyAsync(
-                    new Supplier<>() {
-                        @Override
-                        public String get() {
-                            COUNT.incrementAndGet();
-                            return "Hello, world";
-                        }
-                    },
-                    CompletableFuture.delayedExecutor(5, TimeUnit.SECONDS));
+            return Uni.createFrom().item("Hello, world").onItem().delayIt().by(Duration.ofSeconds(5)).onItem().invoke(
+                    COUNT::incrementAndGet);
+        }
+
+        @GET
+        @Produces(MediaType.TEXT_PLAIN)
+        @Cancellable(false)
+        @Path("2")
+        public Uni<String> uncancellableHello() {
+            COUNT.incrementAndGet();
+            return Uni.createFrom().item("Hello, world").onItem().delayIt().by(Duration.ofSeconds(5)).onItem().invoke(
+                    COUNT::incrementAndGet);
+        }
+
+        @GET
+        @Produces(MediaType.TEXT_PLAIN)
+        @Cancellable
+        @Path("3")
+        public Uni<String> cancellableHello() {
+            COUNT.incrementAndGet();
+            return Uni.createFrom().item("Hello, world").onItem().delayIt().by(Duration.ofSeconds(5)).onItem().invoke(
+                    COUNT::incrementAndGet);
         }
     }
 }
